% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmiss.R
\docType{package}
\name{PPMiss.Package}
\alias{PPMiss.Package}
\alias{_PACKAGE}
\alias{PPMis}
\title{PPMiss: Copula-Based Estimator for Long-Range Dependent Processes under Missing Data.}
\description{
The PPMiss package provides functions for parameter estimation of univariate
time series by means of a parametric copula approach. The methodology is based
on the relationship between the process' covariance decay and a parametric
bivariate copulas associated to lagged variables.
}
\author{
Taiane Schaedler Prass \email{taianeprass@gmail.com} and Guilherme Pumi \email{guipumi@gmail.com}
}
\keyword{internal}
